Make mapping consistent with other mappings.
