+++ api/src/main/java/org/pmiops/workbench/utils/mappers/CommonMappers.java
@@ -1,6 +1,5 @@
 package org.pmiops.workbench.utils.mappers;
 
-import static java.time.temporal.ChronoUnit.DAYS;
 import static org.pmiops.workbench.utils.BillingUtils.isInitialCredits;
 
 import com.google.common.base.Strings;
@@ -9,7 +8,6 @@ import jakarta.inject.Provider;
 import java.sql.Date;
 import java.sql.Timestamp;
 import java.time.Clock;
-import java.time.Instant;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
 import java.util.Optional;
@@ -19,7 +17,6 @@ import org.pmiops.workbench.api.Etags;
 import org.pmiops.workbench.config.WorkbenchConfig;
 import org.pmiops.workbench.db.model.DbCdrVersion;
 import org.pmiops.workbench.db.model.DbUser;
-import org.pmiops.workbench.db.model.DbUserInitialCreditsExpiration;
 import org.pmiops.workbench.db.model.DbWorkspace;
 import org.pmiops.workbench.initialcredits.InitialCreditsService;
 import org.pmiops.workbench.model.BillingStatus;
@@ -154,24 +151,8 @@ public class CommonMappers {
   }
 
   @Named("checkInitialCreditsExtensionEligibility")
-  public boolean checkInitialCreditsExtensionEligibility(DbUser dbUser) {
-    DbUserInitialCreditsExpiration initialCreditsExpiration =
-        dbUser.getUserInitialCreditsExpiration();
-    Instant now = Instant.now();
-    WorkbenchConfig.BillingConfig billingConfig = workbenchConfigProvider.get().billing;
-
-    return initialCreditsExpiration != null
-        && initialCreditsExpiration.getExtensionTime() == null
-        && initialCreditsExpiration.getCreditStartTime() != null
-        && now.isAfter(
-            initialCreditsExpiration
-                .getExpirationTime()
-                .toInstant()
-                .minus(billingConfig.initialCreditsExpirationWarningDays, DAYS))
-        && now.isBefore(
-            initialCreditsExpiration
-                .getCreditStartTime()
-                .toInstant()
-                .plus(billingConfig.initialCreditsExtensionPeriodDays, DAYS));
+  public boolean checkInitialCreditsExtensionEligibility(
+      DbUser dbUser, @Context InitialCreditsService initialCreditsService) {
+    return initialCreditsService.checkInitialCreditsExtensionEligibility(dbUser);
   }
 }
